+++ src/Indice.Features.Cases.App/src/app/features/notifications/notifications.component.ts
@@ -1,9 +1,10 @@
+import { CaseTypePartial } from './../../core/services/cases-api.service';
 import { Component, OnInit } from "@angular/core";
 import { AuthService } from "@indice/ng-auth";
 import { ToasterService, ToastType } from "@indice/ng-components";
 import { EMPTY, forkJoin } from "rxjs";
 import { tap, catchError } from "rxjs/operators";
-import { NotificationSubscriptionViewModel } from "src/app/core/models/NotificationSubscriptionsViewModel";
+import { DisplayNotificationSubscriptionsViewModel, NotificationSubscriptionCategoryViewModel, NotificationSubscriptionViewModel } from "src/app/core/models/NotificationSubscriptionsViewModel";
 import { CasesApiService, NotificationSubscription, NotificationSubscriptionRequest } from "src/app/core/services/cases-api.service";
 
 @Component({
@@ -12,14 +13,17 @@ import { CasesApiService, NotificationSubscription, NotificationSubscriptionRequ
     styleUrls: ['./notifications.component.scss']
 })
 export class NotificationsComponent implements OnInit {
-    public notificationSubscriptionViewModels: NotificationSubscriptionViewModel[] = [];
+    public displayNotificationSubscriptionsViewModel: DisplayNotificationSubscriptionsViewModel | undefined;
     public formSubmitting: boolean = false;
     public loading: boolean = false;
     public isAdmin: boolean = false;
+    public noCategoryName: string = 'ΛΟΙΠΕΣ';
 
-    constructor(private _api: CasesApiService,
+    constructor(
+        private _api: CasesApiService,
         private authService: AuthService,
-        private _toaster: ToasterService) { }
+        private _toaster: ToasterService
+    ) { }
 
     public ngOnInit(): void {
         // awful hack due to @indice/ng-auth's weird behavior
@@ -30,27 +34,18 @@ export class NotificationsComponent implements OnInit {
             getCaseTypes: this._api.getCaseTypes()
         })
             .subscribe(({ getMySubscriptions: mySubscriptions, getCaseTypes: caseTypes }) => {
+                // create an initial view model that contains all categories and case types (and respects server-side ordering) with no active subs
+                this.displayNotificationSubscriptionsViewModel = this.createEmptyDisplayNotificationSubscriptionsViewModel(caseTypes.items!);
                 // add active subscriptions
-                mySubscriptions.notificationSubscriptions?.forEach(sub => {
-                    this.notificationSubscriptionViewModels?.push(new NotificationSubscriptionViewModel(sub, true));
-                });
-                // add inactive subscriptions
-                caseTypes.items?.forEach(caseType => {
-                    let subscription = this.notificationSubscriptionViewModels?.find(x => x.notificationSubscription?.caseTypeId === caseType.id)
-                    if (!subscription) {
-                        subscription = new NotificationSubscriptionViewModel(new NotificationSubscription({ caseTypeId: caseType.id }), false);
-                        this.notificationSubscriptionViewModels?.push(subscription);
-                    }
-                    subscription.title = caseType.title;
-                });
+                this.addActiveSubscriptions(mySubscriptions.notificationSubscriptions!);
                 this.loading = false;
             });
     }
 
     public onSubmit(): void {
         this.formSubmitting = true;
-        let caseTypeIds: string[] = this.notificationSubscriptionViewModels?.filter(x => x.subscribed).map(x => x.notificationSubscription?.caseTypeId!)
-        this._api.subscribe(undefined, new NotificationSubscriptionRequest({ caseTypeIds: caseTypeIds })).pipe(
+        const request = this.createNotificationSubscriptionRequest();
+        this._api.subscribe(undefined, request).pipe(
             tap(_ => {
                 this.formSubmitting = false;
                 this._toaster.show(ToastType.Success, 'Επιτυχής αποθήκευση', `Οι ρυθμίσεις σας αποθηκεύτηκαν επιτυχώς.`, 5000);
@@ -61,6 +56,46 @@ export class NotificationsComponent implements OnInit {
                 return EMPTY;
             })
         ).subscribe();
+    }
+
+    private createNotificationSubscriptionRequest() {
+        let caseTypeIds: string[] = [];
+        this.displayNotificationSubscriptionsViewModel?.categories?.forEach(c =>
+            c.notificationSubscriptions?.filter(x => x.subscribed).forEach(n =>
+                caseTypeIds.push(n.notificationSubscription!.caseTypeId!)));
+        return new NotificationSubscriptionRequest({ caseTypeIds: caseTypeIds })
+    }
+
+    private addActiveSubscriptions(notificationSubscriptions: NotificationSubscription[]) {
+        notificationSubscriptions?.forEach(activeSubscription => {
+            this.displayNotificationSubscriptionsViewModel?.categories?.forEach(category => {
+                category.notificationSubscriptions?.forEach(n => {
+                    if (n.notificationSubscription?.caseTypeId === activeSubscription.caseTypeId) {
+                        n.subscribed = true;
+                        n.notificationSubscription = activeSubscription;
+                    }
+                })
+            })
+        });
+    }
+
+    private createEmptyDisplayNotificationSubscriptionsViewModel(caseTypes: CaseTypePartial[]): DisplayNotificationSubscriptionsViewModel {
+        const categoriesMap = new Map<string, NotificationSubscriptionCategoryViewModel>();
+        for (const caseType of caseTypes) {
+            this.addCaseTypeToCategory(categoriesMap, caseType);
+        }
+        return { categories: Array.from(categoriesMap.values()) };
+    }
+
+    private addCaseTypeToCategory(categoriesMap: Map<string, NotificationSubscriptionCategoryViewModel>, caseType: CaseTypePartial) {
+        let categoryName = caseType?.category?.name ?? this.noCategoryName
+        if (!categoriesMap.has(categoryName)) {
+            categoriesMap.set(
+                categoryName,
+                { name: categoryName, notificationSubscriptions: [] }
+            );
+        }
+        categoriesMap.get(categoryName)!.notificationSubscriptions!.push(new NotificationSubscriptionViewModel(new NotificationSubscription({ caseTypeId: caseType.id }), false, caseType.title!));
     }
 
 }