+++ src/Indice.Features.Cases.UI/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Notification subscriptions now use case type Categories & Ordering
+
 ## [7.0.3 - 2023-03-27]
 ### Added
 - layoutNode option `case-channel` that indicates whether the field is visible based on case's channel.